+++ shoreline.go
@@ -38,7 +38,6 @@ import (
 	"github.com/tidepool-org/go-common/clients/disc"
 	"github.com/tidepool-org/go-common/clients/hakken"
 	"github.com/tidepool-org/go-common/clients/mongo"
-	"github.com/tidepool-org/shoreline/oauth2"
 	"github.com/tidepool-org/shoreline/user"
 	"github.com/tidepool-org/shoreline/user/marketo"
 )
@@ -57,7 +56,6 @@ type (
 		Service disc.ServiceListing `json:"service"`
 		Mongo   mongo.Config        `json:"mongo"`
 		User    user.ApiConfig      `json:"user"`
-		Oauth2  oauth2.ApiConfig    `json:"oauth2"`
 	}
 )
 
@@ -203,8 +201,14 @@ func main() {
 		}
 	}
 
-	clientStore := user.NewMongoStoreClient(&config.Mongo)
-	userapi := user.InitApi(config.User, logger, clientStore, auditLogger, marketoManager)
+	storage, err := user.NewStore(&config.Mongo, logger)
+	if err != nil {
+		logger.Fatal(err)
+	}
+	defer storage.Close()
+	storage.Start()
+
+	userapi := user.InitApi(config.User, logger, storage, auditLogger, marketoManager)
 	logger.Print("installing handlers")
 	userapi.SetHandlers("", rtr)
 
@@ -218,17 +222,6 @@ func main() {
 		Build()
 
 	userapi.AttachPerms(permsClient)
-
-	/*
-	 * Oauth setup
-	 */
-
-	oauthapi := oauth2.InitApi(config.Oauth2, oauth2.NewOAuthStorage(&config.Mongo), userClient, permsClient)
-	oauthapi.SetHandlers("", rtr)
-
-	oauthClient := oauth2.NewOAuth2Client(oauthapi)
-
-	userapi.AttachOauth(oauthClient)
 
 	/*
 	 * Serve it up and publish
@@ -265,6 +258,7 @@ func main() {
 			logger.Printf("Got signal [%s]", sig)
 
 			if sig == syscall.SIGINT || sig == syscall.SIGTERM {
+				storage.Close()
 				server.Close()
 				done <- true
 			}