+++ user/api.go
@@ -19,8 +19,6 @@ import (
 
 	"github.com/tidepool-org/go-common/clients"
 	"github.com/tidepool-org/go-common/clients/status"
-	"github.com/tidepool-org/shoreline/common"
-	"github.com/tidepool-org/shoreline/oauth2"
 	"github.com/tidepool-org/shoreline/user/mailchimp"
 	"github.com/tidepool-org/shoreline/user/marketo"
 
@@ -149,12 +147,11 @@ type (
 		Store            Storage
 		ApiConfig        ApiConfig
 		perms            clients.Gatekeeper
-		oauth            oauth2.Client
 		logger           *log.Logger
 		auditLogger      *log.Logger
 		mailchimpManager mailchimp.Manager
 		loginLimiter     LoginLimiter
-		marketoManager marketo.Manager
+		marketoManager   marketo.Manager
 	}
 	Secret struct {
 		Secret string `json:"secret"`
@@ -183,8 +180,8 @@ type (
 		// Block users to do multiple parallel logins (for load tests we desactivate this)
 		BlockParallelLogin bool `json:blockParallelLogin`
 		//allows for the skipping of verification for testing
-		VerificationSecret string `json:"verificationSecret"`
-		ClinicDemoUserID   string `json:"clinicDemoUserId"`
+		VerificationSecret string           `json:"verificationSecret"`
+		ClinicDemoUserID   string           `json:"clinicDemoUserId"`
 		Mailchimp          mailchimp.Config `json:"mailchimp"`
 		// to create type/file
 		Marketo marketo.Config `json:"marketo"`
@@ -239,7 +236,7 @@ const (
 	STATUS_NO_EXPECTED_PWD       = "No expected password is found"
 )
 
-func InitApi(cfg ApiConfig, logger *log.Logger, store Storage, auditLogger *log.Logger,manager marketo.Manager) *Api {
+func InitApi(cfg ApiConfig, logger *log.Logger, store Storage, auditLogger *log.Logger, manager marketo.Manager) *Api {
 
 	mailchimpManager, err := mailchimp.NewManager(logger, &http.Client{Timeout: 15 * time.Second}, &cfg.Mailchimp)
 	if err != nil {
@@ -259,7 +256,7 @@ func InitApi(cfg ApiConfig, logger *log.Logger, store Storage, auditLogger *log.
 		logger:           logger,
 		auditLogger:      auditLogger,
 		mailchimpManager: mailchimpManager,
-		marketoManager: manager,
+		marketoManager:   manager,
 	}
 
 	api.loginLimiter.usersInProgress = list.New()
@@ -269,10 +266,6 @@ func InitApi(cfg ApiConfig, logger *log.Logger, store Storage, auditLogger *log.
 
 func (a *Api) AttachPerms(perms clients.Gatekeeper) {
 	a.perms = perms
-}
-
-func (a *Api) AttachOauth(client oauth2.Client) {
-	a.oauth = client
 }
 
 func (a *Api) SetHandlers(prefix string, rtr *mux.Router) {
@@ -295,8 +288,6 @@ func (a *Api) SetHandlers(prefix string, rtr *mux.Router) {
 	rtr.HandleFunc("/login", a.Login).Methods("POST")
 	rtr.HandleFunc("/login", a.RefreshSession).Methods("GET")
 	rtr.Handle("/login/{longtermkey}", varsHandler(a.LongtermLogin)).Methods("POST")
-
-	rtr.HandleFunc("/oauthlogin", a.oauth2Login).Methods("POST")
 
 	rtr.HandleFunc("/serverlogin", a.ServerLogin).Methods("POST")
 
@@ -865,72 +856,6 @@ func (a *Api) ServerLogin(res http.ResponseWriter, req *http.Request) {
 	// If the password given at the door is wrong, we cannot generate the token
 	a.logger.Println(http.StatusUnauthorized, STATUS_PW_WRONG)
 	sendModelAsResWithStatus(res, status.NewStatus(http.StatusUnauthorized, STATUS_PW_WRONG), http.StatusUnauthorized)
-	return
-}
-
-// @Summary Login oauth2
-// @Description Login oauth2
-// @ID shoreline-user-api-oauth2login
-// @Accept  json
-// @Produce  json
-// @Success 200 {string} string  "generic json format { \"oauthUser\" : fndUsr, \"oauthTarget\" : result[\"authUserId\"] }"
-// @Header 200 {string} x-tidepool-session-token "authentication token"
-// @Failure 503 {string} string ""
-// @Failure 401 {string} string "generic json format { \"error\" : errorMsg }"
-// @Failure 400 {string} string "generic json format { \"error\" : errorMsg }"
-// @Router /oauthlogin [post]
-func (a *Api) oauth2Login(w http.ResponseWriter, r *http.Request) {
-
-	//oauth is not enabled
-	if a.oauth == nil {
-		a.logger.Println(http.StatusServiceUnavailable, "OAuth is not enabled")
-		w.WriteHeader(http.StatusServiceUnavailable)
-		return
-	}
-
-	if ah := r.Header.Get("Authorization"); ah != "" {
-		if len(ah) > 6 && strings.ToUpper(ah[0:6]) == "BEARER" {
-			if auth_token := ah[7:]; auth_token != "" {
-
-				//check the actual token
-				result, err := a.oauth.CheckToken(auth_token)
-				if err != nil || result == nil {
-					a.logger.Println(http.StatusUnauthorized, "oauth2Login error checking token ", err)
-					w.WriteHeader(http.StatusUnauthorized)
-					return
-				}
-
-				//check the corresponding user
-				fndUsr, errUsr := a.Store.FindUser(&User{Id: result["userId"].(string)})
-				if errUsr != nil || fndUsr == nil {
-					a.logger.Println(http.StatusUnauthorized, "oauth2Login error getting user ", errUsr.Error())
-					w.WriteHeader(http.StatusUnauthorized)
-					return
-				}
-
-				//generate token and send the response
-				if sessionToken, err := CreateSessionTokenAndSave(
-					&TokenData{DurationSecs: 0, UserId: result["userId"].(string), IsServer: false},
-					TokenConfig{DurationSecs: a.ApiConfig.TokenDurationSecs, Secret: a.ApiConfig.Secret},
-					a.Store,
-				); err != nil {
-					a.logger.Println(http.StatusUnauthorized, "oauth2Login error creating session token", err.Error())
-					common.OutputJSON(w, http.StatusUnauthorized, map[string]interface{}{"error": "invalid_token"})
-					return
-				} else {
-					//We are redirecting to the app
-					w.Header().Set(TP_SESSION_TOKEN, sessionToken.ID)
-					common.OutputJSON(w, http.StatusOK, map[string]interface{}{"oauthUser": fndUsr, "oauthTarget": result["authUserId"]})
-					return
-				}
-			}
-		}
-		a.logger.Println(http.StatusUnauthorized, STATUS_AUTH_HEADER_INVLAID)
-		common.OutputJSON(w, http.StatusUnauthorized, map[string]interface{}{"error": STATUS_AUTH_HEADER_INVLAID})
-		return
-	}
-	a.logger.Println(http.StatusBadRequest, STATUS_AUTH_HEADER_REQUIRED)
-	common.OutputJSON(w, http.StatusBadRequest, map[string]interface{}{"error": STATUS_AUTH_HEADER_REQUIRED})
 	return
 }