+++ user/mongoStoreClient.go
@@ -6,10 +6,11 @@ import (
 	"regexp"
 	"sort"
 
-	"github.com/globalsign/mgo"
-	"github.com/globalsign/mgo/bson"
-
-	"github.com/tidepool-org/go-common/clients/mongo"
+	goComMgo "github.com/tidepool-org/go-common/clients/mongo"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 const (
@@ -17,71 +18,43 @@ const (
 	TOKENS_COLLECTION = "tokens"
 )
 
-type MongoStoreClient struct {
-	session *mgo.Session
+// Client struct
+type Client struct {
+	*goComMgo.StoreClient
 }
 
-//We implement the interface from user.Storage
-func NewMongoStoreClient(config *mongo.Config) *MongoStoreClient {
-
-	mongoSession, err := mongo.Connect(config)
-	if err != nil {
-		log.Fatalf("Cannot connect to mongo: %v, %v", config, err)
-	}
-
-	return &MongoStoreClient{
-		session: mongoSession,
-	}
+// NewStore creates a new Client
+func NewStore(config *goComMgo.Config, logger *log.Logger) (*Client, error) {
+	client := Client{}
+	store, err := goComMgo.NewStoreClient(config, logger)
+	client.StoreClient = store
+	return &client, err
 }
 
-func mgoUsersCollection(cpy *mgo.Session) *mgo.Collection {
-	return cpy.DB("").C(USERS_COLLECTION)
+func mgoUsersCollection(c *Client) *mongo.Collection {
+	return c.Collection(USERS_COLLECTION)
 }
 
-func mgoTokensCollection(cpy *mgo.Session) *mgo.Collection {
-	return cpy.DB("").C(TOKENS_COLLECTION)
+func mgoTokensCollection(c *Client) *mongo.Collection {
+	return c.Collection(TOKENS_COLLECTION)
 }
 
-func (d MongoStoreClient) Close() {
-	log.Print(USER_API_PREFIX, "Close the session")
-	d.session.Close()
-	return
-}
-
-func (d MongoStoreClient) Ping() error {
-	// do we have a store session
-	cpy := d.session.Copy()
-	defer cpy.Close()
-
-	if err := cpy.Ping(); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (d MongoStoreClient) UpsertUser(user *User) error {
-
-	cpy := d.session.Copy()
-	defer cpy.Close()
-
+func (c *Client) UpsertUser(user *User) error {
 	if user.Roles != nil {
 		sort.Strings(user.Roles)
 	}
-
+	options := options.Update().SetUpsert(true)
+	update := bson.D{{"$set", user}}
 	// if the user already exists we update otherwise we add
-	if _, err := mgoUsersCollection(cpy).Upsert(bson.M{"userid": user.Id}, user); err != nil {
-		return err
-	}
-	return nil
+	_, err := mgoUsersCollection(c).UpdateOne(c.Context, bson.M{"userid": user.Id}, update, options)
+	return err
 }
 
-func (d MongoStoreClient) FindUser(user *User) (result *User, err error) {
+func (c *Client) FindUser(user *User) (result *User, err error) {
 
 	if user.Id != "" {
-		cpy := d.session.Copy()
-		defer cpy.Close()
-
-		if err = mgoUsersCollection(cpy).Find(bson.M{"userid": user.Id}).One(&result); err != nil {
+		opts := options.FindOne()
+		if err = mgoUsersCollection(c).FindOne(c.Context, bson.M{"userid": user.Id}, opts).Decode(&result); err != nil {
 			return result, err
 		}
 	}
@@ -89,19 +62,34 @@ func (d MongoStoreClient) FindUser(user *User) (result *User, err error) {
 	return result, nil
 }
 
-func (d MongoStoreClient) FindUsers(user *User) (results []*User, err error) {
+func (c *Client) findUsers(filter interface{}, noResultMessage string) (results []*User, err error) {
+	cursor, err := mgoUsersCollection(c).Find(c.Context, filter)
+	defer cursor.Close(c.Context)
+	if err != nil {
+		return results, err
+	}
+	err = cursor.All(c.Context, &results)
+	if err != nil {
+		return results, err
+	}
+	if results == nil {
+		log.Print(noResultMessage)
+		results = []*User{}
+	}
+
+	return results, nil
+}
+
+func (c *Client) FindUsers(user *User) (results []*User, err error) {
 
 	fieldsToMatch := []bson.M{}
-	const (
-		MATCH = `^%s$`
-	)
 
 	if user.Id != "" {
 		fieldsToMatch = append(fieldsToMatch, bson.M{"userid": user.Id})
 	}
 	if user.Username != "" {
-		//case insensitive match
-		fieldsToMatch = append(fieldsToMatch, bson.M{"username": bson.M{"$regex": bson.RegEx{fmt.Sprintf(MATCH, regexp.QuoteMeta(user.Username)), "i"}}})
+		regexFilter := primitive.Regex{Pattern: fmt.Sprintf(`^%s$`, regexp.QuoteMeta(user.Username)), Options: "i"}
+		fieldsToMatch = append(fieldsToMatch, bson.M{"username": bson.M{"$regex": regexFilter}})
 	}
 	if len(user.Emails) > 0 {
 		fieldsToMatch = append(fieldsToMatch, bson.M{"emails": bson.M{"$in": user.Emails}})
@@ -110,94 +98,47 @@ func (d MongoStoreClient) FindUsers(user *User) (results []*User, err error) {
 	if len(fieldsToMatch) == 0 {
 		return []*User{}, nil
 	}
-
-	cpy := d.session.Copy()
-	defer cpy.Close()
-
-	if err = mgoUsersCollection(cpy).Find(bson.M{"$or": fieldsToMatch}).All(&results); err != nil {
-		return results, err
-	}
-
-	if results == nil {
-		log.Printf("no users found: query: (Id = %v) OR (Name ~= %v) OR (Emails IN %v)", user.Id, user.Username, user.Emails)
-		results = []*User{}
-	}
-
-	return results, nil
+	noUserMessage := fmt.Sprintf("no users found: query: (Id = %v) OR (Name ~= %v) OR (Emails IN %v)", user.Id, user.Username, user.Emails)
+	return c.findUsers(bson.M{"$or": fieldsToMatch}, noUserMessage)
 }
 
-func (d MongoStoreClient) FindUsersByRole(role string) (results []*User, err error) {
-	cpy := d.session.Copy()
-	defer cpy.Close()
-
-	if err = mgoUsersCollection(cpy).Find(bson.M{"roles": role}).All(&results); err != nil {
-		return results, err
-	}
-
-	if results == nil {
-		log.Printf("no users found: query: role: %v", role)
-		results = []*User{}
-	}
-
-	return results, nil
+func (c *Client) FindUsersByRole(role string) (results []*User, err error) {
+	noUserMessage := fmt.Sprintf("no users found: query: role: %v", role)
+	return c.findUsers(bson.M{"roles": role}, noUserMessage)
 }
 
-func (d MongoStoreClient) FindUsersWithIds(ids []string) (results []*User, err error) {
-	cpy := d.session.Copy()
-	defer cpy.Close()
-
-	if err = mgoUsersCollection(cpy).Find(bson.M{"userid": bson.M{"$in": ids}}).All(&results); err != nil {
-		return results, err
-	}
-
-	if results == nil {
-		log.Printf("no users found: query: id: %v", ids)
-		results = []*User{}
-	}
-
-	return results, nil
+func (c *Client) FindUsersWithIds(ids []string) (results []*User, err error) {
+	noUserMessage := fmt.Sprintf("no users found: query: id: %v", ids)
+	return c.findUsers(bson.M{"userid": bson.M{"$in": ids}}, noUserMessage)
 }
 
-func (d MongoStoreClient) RemoveUser(user *User) (err error) {
-	cpy := d.session.Copy()
-	defer cpy.Close()
-
-	if err = mgoUsersCollection(cpy).Remove(bson.M{"userid": user.Id}); err != nil {
+func (c *Client) RemoveUser(user *User) (err error) {
+	if _, err := mgoUsersCollection(c).DeleteOne(c.Context, bson.M{"userid": user.Id}); err != nil {
 		return err
 	}
 	return nil
 }
 
-func (d MongoStoreClient) AddToken(st *SessionToken) error {
-	cpy := d.session.Copy()
-	defer cpy.Close()
-
-	if _, err := mgoTokensCollection(cpy).Upsert(bson.M{"_id": st.ID}, st); err != nil {
-		return err
-	}
-
-	return nil
+func (c *Client) AddToken(st *SessionToken) error {
+	options := options.Update().SetUpsert(true)
+	update := bson.D{{"$set", st}}
+	// if the user already exists we update otherwise we add
+	_, err := mgoTokensCollection(c).UpdateOne(c.Context, bson.M{"_id": st.ID}, update, options)
+	return err
 }
 
-func (d MongoStoreClient) FindTokenByID(id string) (*SessionToken, error) {
-	cpy := d.session.Copy()
-	defer cpy.Close()
-
+func (c *Client) FindTokenByID(id string) (*SessionToken, error) {
+	opts := options.FindOne()
 	sessionToken := &SessionToken{}
-	if err := mgoTokensCollection(cpy).Find(bson.M{"_id": id}).One(&sessionToken); err != nil {
+	if err := mgoTokensCollection(c).FindOne(c.Context, bson.M{"_id": id}, opts).Decode(sessionToken); err != nil {
 		return nil, err
 	}
-
 	return sessionToken, nil
 }
 
-func (d MongoStoreClient) RemoveTokenByID(id string) (err error) {
-	cpy := d.session.Copy()
-	defer cpy.Close()
-
-	if err = mgoTokensCollection(cpy).Remove(bson.M{"_id": id}); err != nil {
+func (c *Client) RemoveTokenByID(id string) (err error) {
+	if _, err := mgoTokensCollection(c).DeleteOne(c.Context, bson.M{"_id": id}); err != nil {
 		return err
 	}
-
 	return nil
 }